+++ client/src/features/concept/ConceptPage.tsx
@@ -70,11 +70,11 @@ const ConceptPage: React.FC<{ data: IConcept }> = ({ data }) => {
             />
             <TimelineContainer
               key={`${concept.json.id}-timeline`}
-              searchTags={timeline}
+              halLinkConfig={timeline}
               providedHalLinks={data._links}
             />
-            {/* {Object.keys(data._links).includes(locations.searchTag) && (
-                  <Locations halLink={data._links[locations.searchTag]} />
+            {/* {Object.keys(data._links).includes(locations.halLinkName) && (
+                  <Locations halLink={data._links[locations.halLinkName]} />
                 )} */}
             <Row className="my-3">
               <Col xs={12}>